Rmoved outdated tests from Emib.test.js file + Created new tests for TestFooter component
